requirements.txt
Django==1.9
dj-database-url==0.3.0
django-gravatar2==1.4.0
djangorestframework==3.3.2
gunicorn==19.1.1
newrelic==2.28.0.26
psycopg2==2.5.4
python-json-logger==0.0.6
pytz==2014.7
Unipath==1.1
Whoosh==2.7.0
git+https://github.com/django-haystack/django-haystack.git#egg=django-haystack
python-dotenv==0.1.2
django-secure==1.0.1
